build_standalone/pyproject.toml
[project]
name = "stemify-audio-splitter"
version = "0.1.0"
description = "Backend for Stemify - The Audio Splitter"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.3",
    "flask-cors>=4.0.0",
    "tqdm>=4.36",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "demucs>=4.0.1",
    "mir_eval>=0.7",
    "werkzeug>=3.0.0",
    "dora-search>=0.1.12",
    "diffq>=0.2.1",
    "einops>=0.8.0",
    "flake8>=7.0.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.0",
    "julius>=0.2.3",
    "museval>=0.3.1",
    "mypy>=1.8.0",
    "openunmix>=1.2.0",
    "pyyaml>=6.0",
    "submitit>=1.5.0",
    "treetable>=0.2.6",
    "soundfile>=0.12.0",
    "gunicorn>=21.2.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "pydub>=0.25.1",
    "pywebview>=6.1",
]

[project.optional-dependencies]
standalone = ["pywebview>=6.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
